requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
apscheduler==3.10.4
asn1crypto==1.5.1
asyncio==4.0.0
attrs==25.3.0
bcrypt==4.3.0
bitarray==3.6.1
blessings==1.7
cached-property==2.0.1
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
ckzg==2.1.1
click==8.2.1
cloud-sql-python-connector==1.18.4
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
coverage==7.10.6
cryptography==45.0.6
cytoolz==1.0.1
deepdiff==8.6.0
distlib==0.4.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fastapi==0.116.1
filelock==3.19.1
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hexbytes==1.3.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
inspecta==0.1.3
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.4
mypy_extensions==1.1.0
opentelemetry-api==1.36.0
orderly-set==5.5.0
packaging==25.0
parsimonious==0.10.0
passlib==1.7.4
pg8000==1.31.4
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyproject-api==1.9.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-json-logger==3.3.0
python-multipart==0.0.20
pyunormalize==16.0.0
pywin32==311; sys_platform == "win32"
regex==2025.7.34
requests==2.32.5
rlp==4.1.0
rootpath==0.1.1
rsa==4.9.1
scramp==1.4.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlalchemy-orm==1.2.10
starlette==0.47.2
termcolor==3.1.0
toolz==1.0.0
types-requests==2.32.4.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
web3==7.13.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
